chore: restructure example
sdk.start is no longer async, so we can separate tracer initialization
from the app and launch it using node -r ./tracing.js app.js. The
example has been restructured to work this way in this commit.
